servers/mcp-postgres/pyproject.toml
[project]
name = "mcp-postgres"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "mcp[cli]>=1.26.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
]
